+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 * [feature] markdown: now can copy markdown html.
 * [feature] add Chinese SM3 hash tool. 
 * [feature] add Color converter tool.
+* [feature] date parser tool: add timestamp support.
+* [feature] date parser tool: add helper text.
 * [fix] number base: set input limit.
 * [fix] fix copy error.
 * [fix] update windows make config.
+* [fix] fix date parse l10n.